tests: fold the remaining small tests into main.rs
